requirements.txt
# ==============================
# PredictGW — Dependencies
# ==============================

# Core
pyyaml>=6.0
pandas>=2.0
numpy>=1.24

# Device Communication
requests>=2.31
pymodbus>=3.5

# Analytics (Python fallback)
scikit-learn>=1.3
statsmodels>=0.14
scipy>=1.11

# R Integration (optional)
# rpy2>=3.5

# UI
streamlit>=1.30
plotly>=5.18

# Utilities
schedule>=1.2
